requirements.txt
pycryptodome
colorama~=0.4.4

setuptools~=51.1.0